Add scripts for generating and processing diagrams
The Mathematica script is an example of generating Feynman diagams using
FeynArts (http://feynarts.de), and exporting them to LaTeX. You will
probably want to change the CreateTopology and InsertField commands, the
others should be left alone.
The Python script then processes the output from Mathemtica, and prints
it to stdout - you'll probably want to redirect it to a file for
compiling. The LaTeX needs to go via dvi first, don't use pdflatex.